+++ src/main/java/org/ohdsi/webapi/annotation/annotation/AnnotationDto.java
@@ -6,10 +6,10 @@ import org.ohdsi.webapi.annotation.result.ResultDto;
 
 public class AnnotationDto {
 
-  private Long id;
-  private Long subjectId;
-  private Long sampleId;
-  private Long annotationSetId;
+  private int id;
+  private int subjectId;
+  private int sampleId;
+  private int annotationSetId;
   private String name;
   private Set<ResultDto> results = new HashSet();
 
@@ -18,56 +18,56 @@ public class AnnotationDto {
   /**
    * @return the id
    */
-  public Long getId() {
+  public int getId() {
     return id;
   }
 
   /**
    * @param id the id to set
    */
-  public void setId(Long id) {
+  public void setId(int id) {
     this.id = id;
   }
 
   /**
    * @return the subjectId
    */
-  public Long getSubjectId() {
+  public int getSubjectId() {
     return subjectId;
   }
 
   /**
 	 * @param subjectId the subjectId to set
 	 */
-  public void setSubjectId(Long subjectId) {
+  public void setSubjectId(int subjectId) {
     this.subjectId = subjectId;
   }
 
   /**
    * @return the cohortId
    */
-  public Long getsampleId() {
+  public int getsampleId() {
     return sampleId;
   }
 
   /**
-   * @param cohortId the cohortId to set
+   * @param sampleId the cohortId to set
    */
-  public void setsampleId(Long sampleId) {
+  public void setsampleId(int sampleId) {
     this.sampleId = sampleId;
   }
 
   /**
    * @return the setId
    */
-  public Long getannotationSetId() {
+  public int getannotationSetId() {
     return annotationSetId;
   }
 
   /**
    * @param setId the setId to set
    */
-  public void setannotationSetId(Long setId) {
+  public void setannotationSetId(int setId) {
     this.annotationSetId = setId;
   }
